Rename AnalyticCoolingCurves.py to analytic_cooling_curves.py
https://peps.python.org/pep-0008/#package-and-module-names
