+++ src/fusiondls/__init__.py
@@ -7,7 +7,7 @@ except PackageNotFoundError:
 
     __version__ = get_version(root="..", relative_to=__file__)
 
-from .AnalyticCoolingCurves import cooling_curves
+from .analytic_cooling_curves import cooling_curves
 from .DLScommonTools import file_read, file_write, make_arrays
 from .geometry import MagneticGeometry
 from .solver import SimulationInputs, run_dls
