+++ src/fusiondls/analytic_cooling_curves.py
